+++ src/main/java/com/attica/athens/global/auth/filter/JwtFilter.java
@@ -1,6 +1,7 @@
 package com.attica.athens.global.auth.filter;
 
 import static com.attica.athens.global.auth.jwt.Constants.BEARER_;
+import static com.attica.athens.global.auth.jwt.Constants.REQUEST_ATTRIBUTE_NAME;
 
 import com.attica.athens.domain.common.advice.CustomException;
 import com.attica.athens.global.auth.application.AuthService;
@@ -13,6 +14,7 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.Objects;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -20,13 +22,10 @@ import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 import org.springframework.web.filter.OncePerRequestFilter;
 
 @Slf4j
+@RequiredArgsConstructor
 public class JwtFilter extends OncePerRequestFilter {
 
     private final AuthService authService;
-
-    public JwtFilter(AuthService authService) {
-        this.authService = authService;
-    }
 
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
@@ -50,7 +49,7 @@ public class JwtFilter extends OncePerRequestFilter {
             }
             throw new InvalidAuthorizationHeaderException();
         } catch (CustomException e) {
-            request.setAttribute("jwt exception", e);
+            request.setAttribute(REQUEST_ATTRIBUTE_NAME, e);
         }
 
         filterChain.doFilter(request, response);
