+++ src/main/java/com/attica/athens/global/auth/jwt/Constants.java
@@ -17,4 +17,6 @@ public abstract class Constants {
     public static final String AUTHORIZATION = "Authorization";
 
     public static final String BEARER_ = "Bearer ";
+
+    public static final String REQUEST_ATTRIBUTE_NAME = "jwt exception";
 }
